requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
langchain==0.1.14
langchain-community==0.0.30
langchain-google-genai==0.0.6
google-generativeai==0.8.5
faiss-cpu==1.7.4
PyPDF2==3.0.1



